Rework libvirt roles.
After windows w/ crsp initializer was added
it was found that init processes for win and
lin are similar. but we use different naming
for them.
This patch fixes this problem.

"kvm" occurencies were replaced with "libvirt"
to reduce confusion. This change is needed for
future work also."

<==== Mercurial data ====>
Branch: master